+++ CHANGELOG.md
@@ -14,11 +14,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - #283: Support for native PowerShell remoting as new protocol
 - #1850: Minify config xml
 ### Changed
+- #2013: Removed components check
+- #2011: Removed screenshot manager
 - #2010: Redesigned menus
 - #2005: Removed in-app documentation
 - #1777: Cleaned up VisualStudio project structure
 - #1767: Turned about window into a simple popup form
-- #1766: Converted components check page into options page
 - #1690: Replaced GeckoFX (Firefox) with CefSharp (Chromium)
 - #1325: Language resource files cleanup
 ### Fixed